requirements.txt
boto==2.49.0
boto3==1.14.45
botocore==1.17.45
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
community==1.0.0b1
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
Flask==1.1.2
gensim==3.8.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
lightgbm==2.3.1
louvain==0.7.0
MarkupSafe==1.1.1
matplotlib==3.3.0
networkx==2.4
node2vec==0.3.2
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
smart-open==2.1.0
texttable==1.6.2
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
